Tensor class moved to tensor.py utility.
