--- dnet/utils/__init__.py
@@ -1,1 +1,0 @@
-from jax.interpreters.xla import _DeviceArray as Tensor
