+++ dnet/utils/tensor.py
@@ -1,0 +1,1 @@
+from jax.interpreters.xla import DeviceArray as Tensor
